backend_api/pyproject.toml
[tool.poetry]
name = "tastemaker-code"
version = "0.1.0"
description = ""
authors = ["Alexsandr <artamoshkinas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^69.0.3"
drf-yasg = "^1.21.7"
pillow = "^10.2.0"
drf-spectacular = "^0.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
